add changes

+++ LibraryManagementSystem/src/main/java/com/Service/ReaderServiceImpl.java
@@ -1,6 +1,7 @@
 package com.Service;
 
 import java.util.List;
+import java.util.function.Supplier;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,8 +24,8 @@ public class ReaderServiceImpl implements ReadersService{
 	@Override
 	public Readers updateReaderDetails(Readers reader) 
 	{
-		int Id= reader.getId();
-		Readers r1=repo.findById(Id).orElseThrow();
+		int id= reader.getId();
+		Readers r1=repo.findById(id).orElseThrow();
 		r1.setPassword(reader.getPassword());
 		r1.setFirstName(reader.getFirstName());
 		r1.setLastName(reader.getLastName());
@@ -35,17 +36,20 @@ public class ReaderServiceImpl implements ReadersService{
 	}
 
 	@Override
-	public String deleteReader(int id)
-	{
-		repo.deleteById(id);
-		return "deleted";
+	public boolean deleteReader(int id) throws Throwable {
+		Supplier s = () -> new ReaderNotFoundException("reader not found");
+		Readers reader = repo.findById(id).orElseThrow(s);
+		if ( reader !=null) {
+			repo.deleteById(id);
+			return true;
+		}
+		return false;
 	}
-
 	@Override
 	public List<Readers> viewReadersList()
 	{
-		List<Readers> L1= repo.findAll();
-		return L1;
+		List<Readers> l1= repo.findAll();
+		return l1;
 	}
 
 	@Override
@@ -56,12 +60,12 @@ public class ReaderServiceImpl implements ReadersService{
 	}
 
 	@Override
-	public Boolean loginValidate(String Id, String password) throws ReaderNotFoundException 
+	public boolean loginvalidate(String id, String password) throws ReaderNotFoundException 
 	{
 		boolean flag = false;
 		try 
 		{
-			if (Id == null && password == null) 
+			if (id == null && password == null) 
 			{
 				throw new ReaderNotFoundException("User Details cannot be Empty");
 			} else
